modified files

+++ ivy/functional/backends/numpy/experimental/layers.py
@@ -9,7 +9,7 @@ import ivy
 from ivy.functional.ivy.layers import _handle_padding, _get_num_padded_values
 from ivy.functional.backends.numpy.layers import _add_dilations
 from ivy.functional.ivy.experimental.layers import _padding_ceil_mode
-from ivy.func_wrapper import with_supported_dtypes
+from ivy.func_wrapper import with_supported_dtypes, with_unsupported_dtypes
 from . import backend_version
 
 
@@ -862,4 +862,4 @@ def fft2(
         )
     if norm != "backward" and norm != "ortho" and norm != "forward":
         raise ivy.utils.exceptions.IvyError(f"Unrecognized normalization mode {norm}")
-    return np.fft.fft2(x, s, dim, norm)
+    return np.fft.fft2(x, s, dim, norm).astype(np.complex128)
